requirements.txt
# Core stack (install what you use; backtests/live trading typically need all of these)
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
pyyaml>=6.0
matplotlib>=3.7
requests>=2.28
urllib3>=2.6.3
yfinance>=0.2.30
scipy>=1.10
scikit-learn>=1.4
flask>=3.0
flask-cors>=4.0
asgiref>=3.8
uvicorn[standard]>=0.30
